рефакторинг: выносит функцию в отдельный модуль
