+++ src/main.js
@@ -11,7 +11,7 @@ import MoviesModel from "./models/movies";
 
 import {generateFilms} from "./moks/film";
 import {render} from "./utils/methods-for-components";
-import {Grades} from "./utils/utils";
+import {getUserGrade} from "./utils/utils";
 
 const FILMS_COUNT = 20;
 // const FILMS_EXTRA_COUNT = 2;
@@ -19,27 +19,7 @@ const FILMS_COUNT = 20;
 const films = generateFilms(FILMS_COUNT);
 const moviesModel = new MoviesModel();
 moviesModel.setMovies(films);
-
 const watchedFilms = films.filter((film) => film.alreadyWatched);
-
-const getUserGrade = (movies) => {
-  let grade;
-
-  if (movies.length <= 10) {
-    grade = Grades.NOVICE;
-  } else if (movies.length <= 20) {
-    grade = Grades.FAN;
-  } else if (movies.length >= 21) {
-    grade = Grades.MOVIE_BUFF;
-  } else if (movies.length === 0) {
-    grade = ``;
-  }
-
-  return grade;
-};
-const grade = getUserGrade(watchedFilms);
-
-const user = new UserProfile(grade);
 
 // const filmsExtra = generateFilms(FILMS_EXTRA_COUNT);
 
@@ -47,6 +27,8 @@ const elementHeader = document.querySelector(`.header`);
 const elementFooter = document.querySelector(`.footer__statistics`);
 const elementMain = document.querySelector(`.main`);
 
+const grade = getUserGrade(watchedFilms);
+const user = new UserProfile(grade);
 render(elementHeader, user);
 render(elementFooter, new QuantityFilms(films.length));
